LHC installation circular installation dashboard
This was developed at while working at the excellent TS/CSE group at CERN, to display (live straight from the DB at <https://edms.cern.ch/asbuilt/plsql/mtf_dashboard.circular_dashboard_top>) the current progress of the installation of the QRL (cooling system) and magnets.

This was the first time I ever used SVG, and first proof that SVG really is brilliant, and brilliantly simple. A couple hours after my boss asked if we could somehow display the status in a circle rather than a flat dashboard, to get the big picture, the first actual ring of data was shown on screen.