Chisel points
This is an attempt at a visual proof that there is no 3D object which
has a circle, a square and an equilateral triangle as orthogonal
profiles. The renderings show how either one profile has to be an
isosceles triangle rather than equilateral, or rectangular rather than
square.

If it is indeed possible to create a 3D object fitting all the
requirements as stated, please let me know!

Based on a common riddle (thanks Rolf!) to create a single peg to fit
perfectly into three such holes, as in the popular children's game.